feat(vendor/slack): add slack adaptor for message processing
This might look a little confusing but I think it works pretty well:

Instead of having the slackbot library, we write a custom one of our
own. This gives us more freedom regarding how we structure our project
at the downside of having to write a bit of parsing on our own.

How does it work?

A Slack instance is instantiated with the SlackConfig and a
MessageParser object.

slack.listen(handler) is then called to begin listening to messages from Slack
via `@slackbot name`. When a message is received, a bit if
pre-processing is handled by the generic Slack instance. Afterwards, the
message is passed to the specified MessageHandler.

The MessageHandler can be extended to support multiple types of
messaging parsing. Right now, I've adapted it over docopt (effectively
treating the slackbot interface like a POSIX compliant CLI).

If successful, a dict with the parsed message is passed to the handler
specified in the `.listen` call. This handler through a series of if
else statements can figure out what piece of command to execute.

Bugs

There is one bug responses from the bot to not correctly parse @user
section in the Slack message. Not super critical so I've decided to
commit these changes without fixing it.

Next thing to work on

Implement the handler function to deal with the processed message